Added saving/loading, menu, side scenes, fog
You can now enter and exit internal areas pokemon-style.  The save
system for that is also used in the main menu to let you resume the game
now.  Saving is currently permitted in places which cause problems, so
that needs to be restricted.  I imported a plugin for cubemap-based fog,
so objects go transparent in the distance now.  Several assets were
updated, and some new particle effects were added to the plane and the
new cave scene.  There are also two new skyboxes, one a placeholder and
the other WIP.