Added walking SFX, moving ship parts, updated assets
There is now a sound system with UI control implemented, along with
walking and running sound effects.  The ship now has a propeller and
engines which move appropriately, and particle effects for the engines.
The ship model was updated to include these parts as well as a hold.
There was a serious Git problem, and I don't know if this is
functional/complete, but I'm pushing because I'm tired of this struggle.